# lib/api-zod/src/generated/types/authSession.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MtaaReach CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { User } from './user';

export interface AuthSession {
  user: User;
  token: string;
}