# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MtaaReach CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: number;
  email: string;
  firstName: string;
  lastName: string;
  role: UserRole;
  isActive: boolean;
  /** @nullable */
  tenantId?: number | null;
  /** @nullable */
  assignedCountyId?: number | null;
  /** @nullable */
  assignedConstituencyId?: number | null;
  /** @nullable */
  assignedWardId?: number | null;
  /** @nullable */
  assignedVillageId?: number | null;
  createdAt: string;
}